fix: implemented support do long_int and Uint to '#' flag  Changes to be committed: 	modified:   bonus/ft_print_var_bonus.c 	modified:   bonus/ft_put_bonus.c 	modified:   includes/ft_printf_bonus.h 	modified:   main.c

+++ bonus/ft_print_var_bonus.c
@@ -44,8 +44,8 @@ void	print_bonus(t_element *info, va_list ap, int *len, const char *fmt)
 	if (info->type == '#')
 	{
 		if (*fmt == 'x')
-			ft_put_pointer(va_arg(ap, size_t), len, HEX_LW, 16);
+			ft_put_pointer(va_arg(ap, long int), len, HEX_LW, 16);
 		else
-			ft_put_pointer(va_arg(ap, size_t), len, HEX_UP, 16);
+			ft_put_pointer(va_arg(ap, long int), len, HEX_UP, 16);
 	}
 }
